Remove support for choosing policies
Instead of policies we will try to allow users to add/remove features.
